re-specify selector for .tagInfo padding rule
The previous selector had a couple caveats:

- While writing, the idea of consecutive .tagInfo elements were
considered, and no spacing should be applied inbetween them. A bit of
afterthought made me realize that this selector didn't even do that.

- There was no way (not any easy one at least) to specify a .tagInfo
BEFORE a :not(.tagInfo).

This new selector makes it so that the padding only gets applied if any
other element is present in #tagSuggestions -- this includes other
.tagInfo elements, which is fine for readability in the UI.