cleanup `createBooruFormUI()`
The function was starting to be a bit cumbersome to work with.

UI changes, autocomplete-related instructions, and event listeners in
the function are organized into their own categories.

toggleFormAndSettings replaces the old toggleForm; the old one was meant
to decouple the toggle logic for the booru menu, and in hindsight, it's
really isn't that necessary.

The stored `tagSuggestions` is used instead of a freshly-called
`$("#tagSuggestions")`.

Other changes just make the code nice.